# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/user/delete/{id}": {
            "get": {
                "description": "Get user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/user/get_info/{name}": {
            "get": {
                "description": "Get user by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/user/get_list": {
            "post": {
                "description": "Get user list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user list",
                "parameters": [
                    {
                        "description": "query user request message",
                        "name": "queryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/user/update_info": {
            "post": {
                "description": "Update user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "description": "update user request message",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
        "/api/user/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "login user request message",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tokenString",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Create a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "create user request message",
                        "name": "addRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "middlewares.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.AddUserRequest": {
            "type": "object",
            "required": [
                "user_name",
                "user_password"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "maxLength": 1024
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 256
                },
                "user_password": {
                    "type": "string",
                    "maxLength": 256
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "user_name",
                "user_password"
            ],
            "properties": {
                "user_name": {
                    "type": "string",
                    "maxLength": 256
                },
                "user_password": {
                    "type": "string",
                    "maxLength": 256
                }
            }
        },
...
        "models.QueryUserRequest": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "maximum": 256
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 64
                }
            }
        },
...
        "models.UpdateUserRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "maxLength": 1024
                },
                "id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 256
                },
                "user_password": {
                    "type": "string",
                    "maxLength": 256
                },
                "user_role": {
                    "type": "string",
                    "maxLength": 16
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "secret_key": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_role": {
                    "type": "string"
                }
            }
        }
    }
}